moved memory child-components to new file
